rename dir from app -> src
