+++ garden-backend-service/pyproject.toml
@@ -16,7 +16,6 @@ gunicorn = "^21.2.0"
 [tool.poetry.group.develop]
 optional = true
 
-
 [tool.poetry.group.develop.dependencies]
 pytest = "^7.2.0"
 pytest-mock = "^3.10.0"
