+++ garden-backend-service/tests/test_routes.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 from fastapi.testclient import TestClient
 
-from app.main import app
+from src.main import app
 
 client = TestClient(app)
